Split models into packages
